re-implement wrapper class for all shortcuts
